third_party/WebKit/LayoutTests/svg/transforms/transform-boxsize-usecounter-4.html

<!DOCTYPE html>
<title>SVG transform: 'offset-path' property triggers use counter</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
rect {
  offset-path: path('M0,0');
}
</style>
<svg>
  <rect width="100" height="100" fill="green"/>
</svg>
<script>
// From UseCounter.h
const TransformUsesBoxSizeOnSVG = 1842;

test(function() {
  document.querySelector('rect').getBoundingClientRect();
  assert_true(internals.isUseCounted(document, TransformUsesBoxSizeOnSVG));
});
</script>
